Rework command button layout flow
Ask the layout manager to get the layout info only inside the measure / layout pass when we know if the parent has fixed constraints.

Change the signature of the layout manager to accept constraints, and support fixed width / height in the MEDIUM display state.

Add HorizontalAlignment to control what happens when we have more horizontal space than needed, and use that for secondary / popup content layout.