Upgrade to build 262
As part of this upgrade, there were changes done to the integration of ComposeWindow into the Swing layout pipeline - see more details at https://github.com/JetBrains/compose-jb/issues/917#issuecomment-885594016

Now Aurora uses its own mechanism to precompute the size needed for its popup content (in comboboxes and command buttons) to determine the initial size and location of the matching popup. This removes the need to set initial opacity to 0 and the initial size to arbitrarily large value (1000x1000) and then wait for the first measure pass to get the actual popup content size.